Reorganize lowering pass to make it easier to navigate
